+++ docs/fr_FR/changelog.md
@@ -4,6 +4,9 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte
 
+
+- Possibilité d'activer la rotation automatique de la clef api
+
 # 25/09/2020
 
 - Correction de bugs sur le retour d'activation de l'alarme